Update HapVideoDRAFT.md
My users don't find Hap Alpha too useful due to the quality issues of the RGB channels inherent in DXT5 compression. I'd like to reserve this code for BC7 RGBA images so I can add support in the near future. Encoders are still quite slow (6 secs using CUDA on a K5000 for a 1920x1080 image), but getting to the point where it's usable when needed. I think we'd call this Hap Q Alpha maybe?